read & delete

+++ src/main/java/com/example/crud/response/PostView.java
@@ -6,12 +6,12 @@ import lombok.Getter;
 @Getter
 public class PostView {
 
-    private final int id;
+    private final Long id;
     private final String title;
     private final String content;
 
     @Builder
-    public PostView(int id, String title, String content) {
+    public PostView(Long id, String title, String content) {
         this.id = id;
         this.title = title;
         this.content = content;
